Instance cloning and freezing basic support
migrate is required

Course instance is cloned by using the Save As button in the admin interface
This will copy the instance itself (must be renamed) and any embedded links
from all content graph members. Content graph items are not cloned at this stage.

User answers and stats etc. are now correctly categorised by instance.

Course instances can be frozen (but only from manage.py shell, still figuring
out where in the UI to put the function). Freezing clones the content graph items
associated with the instance and sets the revision of each from None to the
revision id of the latest version. Embedded Links are also treated similarly.

Currently the following things will correctly get the frozen version:
 - content body
 - checkbox/radiobutton choices

The rest is ???

As freezing is based on versioning instead of cloning all of the objects,
it does not clutter the admin interface. However, it is only suitable for
archiving, because frozen version cannot be edited*.

Obvious issues that need addressing:
 - media is managed by instance
 - instance included files ditto
 - terms too

Open questions:
 - answering frozen tasks
 - feedback for frozen content

Secondary future features:
 - implement freeze_to support so that course instances can be frozen to
   calendar dates instead of the latest version only

* Technically you can revert the frozen version, change it, save it,
and manually edit the content graph or embedded link to point to the new
revision, and then finally revert the model to the actual most recent version.