Update to File and Repeated Template Exercise backend download views. They now support versioning fully: it is possible to download past version of the files (previously it was only possible to download the latest file regardless of which revision was being viewed)
The solution is bit of a hack because version number is not available in
the form widget context.

Also unified the two download views to use same method of access control.

Similar views in file upload exercises remain unchanged. Mostly because the
exercise admin for them doesn't support showing revisions anyway.