Added FAQ panels for embedded tasks, and framework for adding other panels and using attention arrows.
FAQ panels contain questions and answers regarding the exercise. They can be
brought up from a link, and will be shown in a separate slide panel on the
left side. Teachers can also edit the questions/answers with a form inside
the panel.

FAQ questions can be attached to hooks similar to triggerable highlights.
When triggers are included in JSON evaluation, the user will be notified
about the existence of the FAQ panel with an attention arrow, and if the
panel is opened, any relevant questions will be pre-expanded and highlighted.

FAQ questions are revision-managed.

This update requires migration and collectstatic